Adding solution for problem 19.
This solution seems more complex than it should be since it runs quick and
the problem itself is more or less trivial. I'll blame the complexity on the
weirdness of the calendar system, and the need to deal with that. There's
probably a way to get around the 25 lines of week/month definitions, but it
was the clearest way to deal with things that I could come up with.